Simplify special variable handling, fix id clash
The older code was built onto the assumption that special variables were handled exclusively by certain instructions and that those instructions didn't do anything but handle special variables - which is not true, e.g. oppop. This modifies the index of the variable straight into the structure by swapping VariableDefinitions in the VARI array instead when necessary.